+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.0.60] - 2016-10-07
 ### Added
 - FEATURE kumo now has the visualize cmd. Req. dot installation (#136).
+- FIX moved tests to pytest to improve cleanup after tests (#119).
+- FIX ramuda rollback to previous version.
+- FIX ramuda fail deployment on failing ping (#113).
+- FIX kumo Parameter diffing does not work for aws coma-seperated inputs (#77).
+- FIX kumo Parameter diffing does not work for aws coma-seperated inputs (#77).
 - FEATURE tenkai now has the slack notifications (#79).- FIX moved tests to pytest to improve cleanup after tests (#119).
 
 ## [0.0.57] - 2016-09-23